# Remove dynamic dispatch in Hessian evaluation