As `chunk` is maximum 10, we can reuse the `_create_binary_switch` we use for type-stable evaluation of our list of operators to call `_eval_hessian_inner` in a type-stable way.
On this benchmark https://github.com/jump-dev/MathOptInterface.jl/pull/2739#issuecomment-2833683884
Before this PR
```
5.959426 seconds (158.26 k allocations: 11.273 MiB)
5.973144 seconds (158.26 k allocations: 11.273 MiB)
6.078342 seconds (158.26 k allocations: 11.273 MiB)
6.031799 seconds (158.26 k allocations: 11.273 MiB, 0.23% gc time)
6.022742 seconds (158.26 k allocations: 11.273 MiB)
6.013003 seconds (158.26 k allocations: 11.273 MiB)
6.021324 seconds (158.26 k allocations: 11.273 MiB)
```
After this PR:
```
5.925775 seconds
5.820862 seconds
5.865539 seconds
5.823728 seconds
5.848671 seconds
5.864283 seconds
5.876498 seconds
5.911389 seconds
5.894025 seconds
```