@odow 2025-04-28
I'm very very very hesitant about this. At minimum, I think we should manually write out the switch. Macros and `@eval` are just too complicated and don't justify the extra 20 lines of code. It's a different story when there are 70 something operators.

It would also mean we need tests for every possible chunk size to ensure 100% code coverage.