@odow 2025-04-29
It's nice, there are already tests that check all of the different chunk sizes.