@blegat 2025-04-29
> It would also mean we need tests for every possible chunk size to ensure 100% code coverage.

I don't agree. We're using a binary switch that is already heavily tested, what's the risk ?
If you hardcode 10 different `if-else` then you might make a typo but not with the binary switch.
Also, the advantage of the switch is that you can modify that constant without the need to modify the code