@odow 2025-04-29
I've changed to use the binary switch. I wouldn't say that it is "heavily" tested. It's used in one other place.

> If you hardcode 10 different if-else then you might make a typo but not with the binary switch.

Not if we have every line tested :smile: