@blegat 2025-04-30
> I wouldn't say that it is "heavily" tested. It's used in one other place.

Maybe that was an over-exageration :smile: But if the binary switch had a bug I prefer MOI to be completely broken rather than having a hidden bug. In that sense, giving it even more coverage is good :)